+++ packages/apollo-angular/CHANGELOG.md
@@ -3,9 +3,9 @@
 ### vNEXT
 
 * Introduces `ExtraSubscriptionOptions`. Allow to run `Apollo.subscribe` outside
-  the Zone.
+  the Zone ([PR #488](https://github.com/apollographql/apollo-angular/pull/488))
 * Exposes `wrapWithZone`, in case someone wants to run `Observable` inside the
-  Zone.
+  Zone ([PR #488](https://github.com/apollographql/apollo-angular/pull/488))
 
 ### v1.0.1